Remove rapids_udf_test.py, it should be in and tested in the external repository
