+++ integration_tests/src/main/python/row-based_udf_test.py
@@ -18,7 +18,18 @@ from asserts import assert_gpu_and_cpu_are_equal_sql
 from data_gen import *
 from spark_session import with_spark_session
 from conftest import skip_unless_precommit_tests
-from rapids_udf_test import skip_if_no_hive, load_hive_udf
+
+def drop_udf(spark, udfname):
+    spark.sql("DROP TEMPORARY FUNCTION IF EXISTS {}".format(udfname))
+
+def skip_if_no_hive(spark):
+    if spark.conf.get("spark.sql.catalogImplementation") != "hive":
+        skip_unless_precommit_tests('The Spark session does not have Hive support')
+
+def load_hive_udf(spark, udfname, udfclass):
+    drop_udf(spark, udfname)
+    # if UDF failed to load, throws AnalysisException, check if the udf class is in the class path
+    spark.sql("CREATE TEMPORARY FUNCTION {} AS '{}'".format(udfname, udfclass))
 
 def test_hive_empty_simple_udf():
     with_spark_session(skip_if_no_hive)
